rename: Location -> PresentLocation 파일명 변경
현위치 프로바이더와 클릭한 위치 프로바이더 차별화를 두기 위해 파일명 변경
