+++ src/index.tsx
@@ -3,7 +3,9 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { GlobalStyle } from '@styles/globalStyle';
 import App from './App';
-import { LocationProvider } from '@provider/Location';
+import { LocationProvider } from '@provider/PresentLocation';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { queryClient } from '@api/instance';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
